refactor: ProfileMenu.jsx renamed to *.tsx
