+++ src/components/profile-menu/ProfileMenu.tsx
@@ -1,12 +1,12 @@
-import React, {useEffect} from 'react';
+import React, {useEffect, FC} from 'react';
 import styles from "./profile-menu.module.css";
 import {NavLink, useNavigate} from "react-router-dom";
 import {Outlet} from "react-router";
 import {getLogOutUser, getUserInfo} from "../../services/actions/user";
-import {useDispatch, useSelector} from "react-redux";
+import {useDispatch, useSelector} from '../../services/hooks';
 import {getCookie} from "../../utils/utils";
 
-const ProfileMenu = () => {
+const ProfileMenu: FC = () => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
   const token = getCookie('token');
@@ -25,7 +25,7 @@ const ProfileMenu = () => {
   }, // eslint-disable-next-line
     [accessToken]);
 
-  const setActive = ({isActive}) => isActive ? `${styles.link} text text_type_main-medium ${styles.link_active}`
+  const setActive = ({isActive}: {isActive: boolean}) => isActive ? `${styles.link} text text_type_main-medium ${styles.link_active}`
     : `${styles.link} text text_type_main-medium`;
 
   return (
